requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.47
pydantic==1.10.2
python-dotenv==1.0.0
pytest==7.4.0  
requests==2.31.0  
pymysql==1.0.3

